+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 [Full diff](https://github.com/ybiquitous/aufgaben/compare/0.7.1...HEAD)
 
 - Support Ruby 3.0.2, 2.7.4, and 2.6.8 for `bump:ruby` [#38](https://github.com/ybiquitous/aufgaben/pull/38)
+- Colored message for release [#39](https://github.com/ybiquitous/aufgaben/pull/39)
 
 ## 0.7.1